[clean-up] Move shared functions to utility.py to avoid import loops.
