+++ perlin.py
@@ -4,10 +4,10 @@ import xml.etree.ElementTree as ET
 from typing import List, Tuple
 from xml.etree import ElementTree
 
-from render import toimage
-
 import noise
 import numpy as np
+
+from utility import distance, find_city_centre, radius_of_network
 
 if 'SUMO_HOME' in os.environ:
     tools = os.path.join(os.environ['SUMO_HOME'], 'tools')
@@ -22,10 +22,10 @@ import sumolib
 # POPULATION_BASE = random.randrange(-1000, 1000)
 POPULATION_BASE: float = 0
 # INDUSTRY_BASE = random.randrange(-1000, 1000)
-INDUSTRY_BASE: float = 0
+INDUSTRY_BASE: float = 2
 
 
-def get_edge_pair_centroid(coords: List[Tuple[float, float]]) -> (float, float):
+def get_edge_pair_centroid(coords: List[Tuple[float, float]]) -> Tuple[float, float]:
     """
     Centroid of rectangle (edge_pair) = (width/2, height/2)
     :param coords: [(x_1,y_1), (x_2,y_2), ... , (x_n,y_n)]
@@ -33,7 +33,7 @@ def get_edge_pair_centroid(coords: List[Tuple[float, float]]) -> (float, float):
     """
     x_avg = np.mean([pos[0] for pos in coords])
     y_avg = np.mean([pos[1] for pos in coords])
-    return x_avg, y_avg
+    return float(x_avg), float(y_avg)
 
 
 def get_perlin_noise(x: float, y: float, base: float, scale: float = 0.005, octaves: int = 3) -> float:
@@ -62,7 +62,7 @@ def get_population_number(edge: sumolib.net.edge.Edge, base: float, centre,
     :return: the scaled noise value as float in [0:1]
     """
     x, y = get_edge_pair_centroid(edge.getShape())
-    return get_perlin_noise(x=float(x), y=float(y), base=base, scale=scale, octaves=octaves) + (
+    return get_perlin_noise(x, y, base=base, scale=scale, octaves=octaves) + (
             1 - (distance((x, y), centre) / radius))
 
 
@@ -78,7 +78,6 @@ def apply_network_noise(net: sumolib.net.Net, xml: ElementTree, scale: float = 0
     # Calculate and apply Perlin noise for all edges in network to population in statistics
     print("Writing Perlin noise to population and industry")
 
-    from randomActivityGen import find_city_centre
     centre = find_city_centre(net)
     radius = radius_of_network(net, centre)
 
@@ -103,47 +102,3 @@ def apply_network_noise(net: sumolib.net.Net, xml: ElementTree, scale: float = 0
                 "population": str(population),
                 "workPosition": str(industry)
             })
-
-
-def radius_of_network(net: sumolib.net.Net, centre):
-    """
-    Get distance from centre to outermost node. Use this for computing radius of network.
-    :return: the radius of the network
-    """
-    return np.max([distance(centre, node.getCoord()) for node in net.getNodes()])
-
-
-def distance(pos1, pos2):
-    """
-    Return the distance between two points in a plane
-    :return: the distance between pos1 and pos2
-    """
-    x1, y1 = pos1
-    x2, y2 = pos2
-    return np.sqrt((x2 - x1) ** 2 + (y2 - y1) ** 2)
-
-
-def display_noisemap(net: sumolib.net.Net, scale: float = 0.005, octave: int = 3):
-    """
-    Draw an image of the noisemap applied to a given network. Currently only displaying residential densities.
-    FIXME: do residential and industrial in different colours
-    :param net: the network to display noisemap for
-    :param scale: the scale to multiply to each coordinate, default is 0.005
-    :param octave: the octaves to use when sampling, default is 3
-    :return:
-    """
-    boundary = net.getBoundary()
-    size = (boundary[2], boundary[3])
-    from randomActivityGen import find_city_centre
-    centre = find_city_centre(net)
-    radius = radius_of_network(net, centre)
-
-    # Initialise noisemap
-    arr = [[0 for _ in range(int(size[0]))] for _ in range(int(size[1]))]
-
-    for i in range(0, int(size[0])):
-        for j in range(0, int(size[1])):
-            p_noise = get_perlin_noise(i, j, base=POPULATION_BASE, scale=scale, octaves=octave)
-            arr[j][i] = p_noise + (1 - (distance((i, j), centre) / radius))
-
-    toimage(arr).show()
