+++ utility.py
@@ -1,0 +1,38 @@
+import os
+import sys
+import numpy as np
+from typing import Tuple
+
+if 'SUMO_HOME' in os.environ:
+    tools = os.path.join(os.environ['SUMO_HOME'], 'tools')
+    sys.path.append(tools)
+else:
+    sys.exit("Please declare environment variable 'SUMO_HOME' to use sumolib")
+
+import sumolib
+
+
+def find_city_centre(net: sumolib.net.Net) -> Tuple[float, float]:
+    """
+    Finds the city centre; average node coord of all nodes in the net
+    """
+    node_coords = [node.getCoord() for node in net.getNodes()]
+    return float(np.mean([c[0] for c in node_coords])), float(np.mean([c[1] for c in node_coords]))
+
+
+def radius_of_network(net: sumolib.net.Net, centre: Tuple[float, float]):
+    """
+    Get distance from centre to outermost node. Use this for computing radius of network.
+    :return: the radius of the network
+    """
+    return np.max([distance(centre, node.getCoord()) for node in net.getNodes()])
+
+
+def distance(pos1: Tuple[float, float], pos2: Tuple[float, float]):
+    """
+    Return the distance between two points in a plane
+    :return: the distance between pos1 and pos2
+    """
+    x1, y1 = pos1
+    x2, y2 = pos2
+    return np.sqrt((x2 - x1) ** 2 + (y2 - y1) ** 2)
